% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_functions.R
\name{beach_get_data}
\alias{beach_get_data}
\title{Retrieve data for danish beaches}
\usage{
beach_get_data()
}
\value{
a list of tibbles with the data
}
\description{
Retrieve data for danish beaches
}
\examples{
beach_data <- badevand::beach_get_data()

}
