% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_functions.R
\name{extract_links}
\alias{extract_links}
\title{Extract the links}
\usage{
extract_links(call_content)
}
\arguments{
\item{call_content}{the data from the api call}
}
\value{
a tibble
}
\description{
Extracts the links where available
}
