% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_functions.R
\name{extract_meassurements}
\alias{extract_meassurements}
\title{Extract the meassurements}
\usage{
extract_meassurements(call_content)
}
\arguments{
\item{call_content}{the data from the api call}
}
\value{
a tibble
}
\description{
Extract the meassurements
}
