% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_functions.R
\name{extract_overview_table}
\alias{extract_overview_table}
\title{Extract overview of beaches}
\usage{
extract_overview_table(call_content)
}
\arguments{
\item{call_content}{the data from the api call}
}
\value{
a tibble
}
\description{
Extract overview of beaches
}
