% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the_functions.R
\name{retrieve_api_data}
\alias{retrieve_api_data}
\title{Retrieve data from the api}
\usage{
retrieve_api_data()
}
\value{
a list of data
}
\description{
Calls the API, retrieves the data and returns it as a list.
}
