% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_quality}
\alias{water_quality}
\title{Lookup table for water quality.}
\format{A tibble with 4 rows and 3 variables:
\describe{
  \item{value}{the value corresponding to the water_quality column in the API data}
  \item{description}{the description of the water quality}
  \item{icon}{a link to a relevant icon for the water quality in question}
}}
\source{
\url{http://api.vandudsigten.dk/doc/beaches}
}
\usage{
water_quality
}
\description{
A tibble containing the description of the water quality and a
link to the relevant icon.
}
\keyword{datasets}
